refactor(fe): fetchHint 중복 코드 제거 및 더블클릭 guard 추가
